scanner/requirements.txt
python-binance==1.0.19
asyncpg==0.29.0
aiohttp==3.9.3
pandas==2.2.0
numpy==1.26.4
asyncio-throttle==1.0.2
